fix: Split Docker tests for CI compatibility
- Move Docker runtime tests to test_docker.py (skipped in CI)
- Create test_docker_basic.py with file validation tests (run in CI)
- Skip Docker-in-Docker tests when CI or GITHUB_ACTIONS env vars are set
- Fix CI test failures by separating static vs runtime Docker tests

The runtime Docker tests require actual Docker daemon access and are
automatically skipped in CI environments. File validation tests always run.
