+++ tests/test_docker.py
@@ -1,4 +1,9 @@
-"""Tests for Docker container functionality"""
+"""Tests for Docker container functionality
+
+These tests require Docker to be installed and running.
+They are skipped in CI environments where Docker-in-Docker is not available.
+Run these tests locally with: pytest tests/test_docker.py
+"""
 
 import os
 import subprocess
@@ -8,10 +13,21 @@ import pytest
 
 
 def docker_available():
-    """Check if Docker is available"""
+    """Check if Docker is available and usable"""
+    # Skip Docker tests in CI environment (act or GitHub Actions)
+    if os.environ.get("CI") == "true" or os.environ.get("GITHUB_ACTIONS") == "true":
+        return False
+
     try:
         result = subprocess.run(
             ["docker", "--version"], capture_output=True, text=True, check=False
+        )
+        if result.returncode != 0:
+            return False
+
+        # Also check if Docker daemon is running
+        result = subprocess.run(
+            ["docker", "info"], capture_output=True, text=True, check=False
         )
         return result.returncode == 0
     except FileNotFoundError:
@@ -351,54 +367,5 @@ class TestDockerCompose:
         assert result.returncode == 0 or len(result.stderr) == 0
 
 
-def test_dockerfile_exists():
-    """Test that Dockerfile exists"""
-    assert os.path.exists("Dockerfile")
-
-
-def test_dockerignore_exists():
-    """Test that .dockerignore exists"""
-    assert os.path.exists(".dockerignore")
-
-
-def test_docker_compose_exists():
-    """Test that docker-compose.yml exists"""
-    assert os.path.exists("docker-compose.yml")
-
-
-def test_dockerfile_syntax():
-    """Test basic Dockerfile syntax"""
-    with open("Dockerfile", "r") as f:
-        content = f.read()
-
-    # Check for required elements
-    assert "FROM python:3.11-slim" in content
-    assert "WORKDIR /app" in content
-    assert "USER mcp" in content  # Non-root user
-    assert "HEALTHCHECK" in content
-    assert "FAL_KEY" in content
-    assert "EXPOSE 8080" in content
-
-
-def test_dockerignore_content():
-    """Test .dockerignore has proper exclusions"""
-    with open(".dockerignore", "r") as f:
-        content = f.read()
-
-    # Check for common exclusions
-    assert ".git" in content
-    assert "__pycache__" in content
-    assert ".env" in content
-    assert "*.pyc" in content or "*.py[cod]" in content
-
-
-def test_docker_compose_services():
-    """Test docker-compose.yml has proper services"""
-    with open("docker-compose.yml", "r") as f:
-        content = f.read()
-
-    # Check for required services
-    assert "fal-mcp-http" in content
-    assert "FAL_KEY=${FAL_KEY}" in content
-    assert "8080:8080" in content
-    assert "healthcheck" in content
+# File validation tests moved to test_docker_basic.py
+# These runtime tests require actual Docker daemon access
